Core: silnced warning in heaputil.cpp #748
git-svn-id: svn://ultimatepp.org/upp/trunk@7258 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/heaputil.cpp
@@ -94,7 +94,7 @@ void *SysAllocRaw(size_t size, size_t reqsize)
 	}
 	if(!ptr)
 		OutOfMemoryPanic(size/*reqsize*/);
-	sKB += rsz;
+	sKB += (int)rsz;
 	DoPeakProfile();
 	return ptr;
 }
